// frontend/src/app/models/pageResponse.model.ts
export interface PageResponse<T> {
    content: T[];
    pageable: {
        pageNumber: number;
        pageSize: number;
        sort: {
            empty: boolean;
            sorted: boolean;
            unsorted: boolean;
        };
        offset: number;
        paged: boolean;
        unpaged: boolean;
    };
    totalPages: number;
    totalElements: number;
    first: boolean;
    last: boolean;
    number: number;
    size: number;
    numberOfElements: number;
    empty: boolean;
}